+++ src/main/java/com/vigilante/retriever/v1/drug/application/query/DrugMongoQuery.java
@@ -22,4 +22,8 @@ public class DrugMongoQuery {
 	public DrugEntity getById(String id) {
 		return drugMongoPort.findById(id).orElseThrow(DrugNotFoundException::new);
 	}
+
+	List<DrugEntity> getByArgot(String argot) {
+		return drugMongoPort.findByArgot(argot);
+	}
 }
