+++ src/main/java/com/vigilante/retriever/v1/drug/domain/entity/DrugEntity.java
@@ -1,12 +1,22 @@
 package com.vigilante.retriever.v1.drug.domain.entity;
 
+import java.util.List;
+
 import lombok.Builder;
 
 @Builder
 public record DrugEntity(
-	String _id,
-	String drugName,
+	String id,
+	String drugBankId,
+	String name,
 	String drugType,
-	String drugEnName
+	String EnglishName,
+	List<Argot> argots
 ) {
+	@Builder
+	public record Argot(
+		String name,
+		String description
+	) {
+	}
 }
