+++ src/main/java/com/vigilante/retriever/v1/drug/domain/exception/ArgotNotFoundException.java
@@ -1,6 +1,6 @@
-package com.vigilante.retriever.v1.argot.domain.exception;
+package com.vigilante.retriever.v1.drug.domain.exception;
 
-import static com.vigilante.retriever.v1.argot.domain.code.ArgotErrorCode.*;
+import static com.vigilante.retriever.v1.drug.domain.code.ArgotErrorCode.*;
 
 import com.vigilante.retriever.common.domain.exception.NotFoundException;
